+++ Sources/LazyInjector.swift
@@ -24,6 +24,18 @@ public struct LazyInjector<K: ProvidableKey>: InjectorDerivingFromMutableInjecto
                                                     usingFactory: factory)
     }
 
+    #if swift(>=3.0)
+    /// See `MutableInjector.revoke(key:)`.
+    public mutating func revoke(key: K) {
+        lazyProviders.removeValue(forKey: key)
+    }
+    #else
+    /// See `MutableInjector.revoke(key:)`.
+    public mutating func revoke(key key: K) {
+        lazyProviders.removeValueForKey(key)
+    }
+    #endif
+
     /// See `Injector.providedKeys`.
     public var providedKeys: [K] {
         return Array(lazyProviders.keys)